Vertices to curve tool
Ideally used for creating a variety of motion paths sourcing a given mesh. The curve is created from a selection of vertices, pertaining to the order that the verts are selected in. Then you can attach a object to the curve and have it move along the path oriented according to a given mesh's normals.